+++ src/Microsoft.OpenTelemetry/Agent365/Runtime/DTOs/Builders/ExecuteToolDataBuilder.cs
@@ -218,8 +218,16 @@ namespace Microsoft.Agents.A365.Observability.Runtime.DTOs.Builders
             if (transferDetails != null)
             {
                 AddIfNotNull(attributes, OpenTelemetryConstants.TransferModeKey, transferDetails.ModeValue);
-                AddIfNotNull(attributes, OpenTelemetryConstants.TransferTargetNameKey, transferDetails.TargetName);
-                AddIfNotNull(attributes, OpenTelemetryConstants.TransferTargetTypeKey, transferDetails.TargetTypeValue);
+
+                var targetAgentDetails = transferDetails.TargetAgentDetails;
+                if (targetAgentDetails != null)
+                {
+                    AddIfNotNull(attributes, OpenTelemetryConstants.TransferTargetAgentIdKey, targetAgentDetails.AgentId);
+                    AddIfNotNull(attributes, OpenTelemetryConstants.TransferTargetAgentNameKey, targetAgentDetails.AgentName);
+                    AddIfNotNull(attributes, OpenTelemetryConstants.TransferTargetAgentBlueprintIdKey, targetAgentDetails.AgentBlueprintId);
+                    AddIfNotNull(attributes, OpenTelemetryConstants.TransferTargetAgentPlatformIdKey, targetAgentDetails.AgentPlatformId);
+                    AddIfNotNull(attributes, OpenTelemetryConstants.TransferTargetAgentVersionKey, targetAgentDetails.AgentVersion);
+                }
             }
         }
     }